apps/pipeline/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
livekit==0.12.3
deepgram-sdk==3.5.0
groq==0.9.0
kokoro-onnx==0.1.3
numpy==1.26.4
soundfile==0.12.1
